change state machine folder structure
