+++ src/state-machine/continue.js
@@ -1,0 +1,7 @@
+function continueRobot (command) {
+  console.log('CONTINUANDO ANÁLISE')
+}
+
+export {
+  continueRobot
+}
