+++ src/state-machine/init.js
@@ -1,0 +1,7 @@
+function initAnalisys (command) {
+  console.log('INICIANDO ANALISE')
+}
+
+export {
+  initAnalisys
+}
