+++ src/state-machine/pause.js
@@ -1,0 +1,7 @@
+function pauseRobot (command) {
+  console.log('PAUSANDO ROBO')
+}
+
+export {
+  pauseRobot
+}
