+++ src/state-machine/reset.js
@@ -1,0 +1,7 @@
+function resetState (command) {
+  console.log('RESETANDO ESTADO DO ROBO')
+}
+
+export {
+  resetState
+}
